+++ src/app/app.html
@@ -1,5 +1,5 @@
 <app-header></app-header>
 
-<app-user-input (calculate)="onCalculateInvestmentResults($event)" />
+<app-user-input />
 
-<app-investment-results [data]="resultsData()"> </app-investment-results>
+<app-investment-results> </app-investment-results>
