+++ src/app/investment-results/investment-results.component.html
@@ -1,4 +1,4 @@
-@if(!data()){
+@if(!results){
 <p>Please enter some values and press "Calculate"</p>
 } @else{
 
@@ -13,7 +13,7 @@
     </tr>
   </thead>
   <tbody>
-    @for(result of data(); track result.year) {
+    @for(result of results; track result.year) {
     <tr>
       <td>{{ result.year }}</td>
       <td>{{ result.valueEndOfYear | currency }}</td>
